regrettably... delete align.h and put copies in bbccon.c and BBC.h I don't imagine these would change much in the future.
